HitboxComponent disable/enable feature
Was looking at ways to have the frog's hitbox only active mid jump, and thought it'd be useful to have a way to toggle the Hitbox's collider via code. I was doing a `.get_node()` style thing, but wanted to make it reliable, so I put the functionality into the HitboxComponent itself. I also wanted to try out a lambda function, which then let me make `.call_deferred()` a part of the method.

call_deferred runs *after* the physics frame, which is necessary when changing colliders. Changing them mid-physics frame causes issues, is my understanding.