Renamed item scenes
I renamed the item scenes to reflect what they're for.
Static has a Sprite2D and only 1 texture.
Animated has an AnimatedSprite2D and has 2 textures that it turns into an animation at runtime.
If we wanted, we could combine them and have an if/else check for assigning it a texture or an animation.