Player hitbox disable/enable, Hurt animation
Added a hurt animation to the player. Might want to move it over to an AnimationPlayer node at some point.

The player now also only has a hitbox when airborne, like the frog. Forgot about this earlier.

Fun move_and_slide bug: when the player walks into the bottom "stair step" of the first/grassy platform,  the frog on the right will push the player down, then up, then die because it pushed the player onto it's own head.