Added Worm, Enemy State Machine
The beginnings of a state machine for enemies, with possible expansion to the player.

Idle causes the enemy to stand still for a random amount of time.
Wander has the enemy pick a direction and move slowly that way for a random duration.
Chase has the enemy move quicker towards the player when the player is nearby during the Wander state.