Add Hit Flash and Death animation to enemies
Enemies and player will flash white when hit.
The flash just applies a shader that turns the sprite pure white, with a tweened fadeout.
The death animation is a particle effect that gets detached from the main body before it's removed from the scene.
Changed the player's bounce on hit to set velocity instead of adding. Managed to get the player to fly a few times the way it was.
Could theoretically add the death component to the player, but the scene pauses on death, so nothing will happen right now.