Frog gets state machine, enemy states expanded, frog sprite added
I had to edit the frog sprites to get one with closed eyes and standing. Why default idle is jumping, I'll never know. I assume it's swimming.
Frog has been converted to the state machine model. I went with adding functionality to the chase state for this update, and have decided that in the future it'd probably be better to make new classes that inherit from the chase state. There's some stuff in there planned for enemies like the saw blade / fly / bee /etc.