Player ItemBox Collider disabled at wrong time fix
Noticed an error I hadn't seen before, might've only occurred sometimes depending on timing.
The player's item box collider was being disabled and enabled depending on if the player was jumping or not / moving up vs down. Before it could be disabled while area collisions are still being processed which throws an error, it should now wait for the idle frame in between collision checks.
Also attaching the collision mask change to the player's hurtbox. It's simply set to defaults instead of overriding.